requirements.txt
streamlit
langchain
langchain-core
langchain-google-genai
python-dotenv
sqlite3-binary
